// src/types/graphql.tsx
export type Maybe<T> = T | null
// ...
export type Scalars = {
  ID: { input: string; output: string }
  String: { input: string; output: string }
  Boolean: { input: boolean; output: boolean }
  Int: { input: number; output: number }
  Float: { input: number; output: number }
  ISO8601DateTime: { input: any; output: any }
  JsonScalar: { input: any; output: any }
}

export type Alternate = {
  __typename?: 'Alternate'
  fullSlug: Scalars['String']['output']
  id: Scalars['Int']['output']
  isFolder?: Maybe<Scalars['Boolean']['output']>
  name: Scalars['String']['output']
  parentId?: Maybe<Scalars['Int']['output']>
  published: Scalars['Boolean']['output']
  slug: Scalars['String']['output']
}

export type Asset = {
  __typename?: 'Asset'
  alt?: Maybe<Scalars['String']['output']>
  copyright?: Maybe<Scalars['String']['output']>
  filename: Scalars['String']['output']
  focus?: Maybe<Scalars['String']['output']>
  id?: Maybe<Scalars['Int']['output']>
  name?: Maybe<Scalars['String']['output']>
  title?: Maybe<Scalars['String']['output']>
}

export type ContentItem = {
  __typename?: 'ContentItem'
  alternates?: Maybe<Array<Maybe<Alternate>>>
  content?: Maybe<Scalars['JsonScalar']['output']>
  content_string?: Maybe<Scalars['String']['output']>
  created_at?: Maybe<Scalars['String']['output']>
  default_full_slug?: Maybe<Scalars['String']['output']>
  first_published_at?: Maybe<Scalars['String']['output']>
  full_slug?: Maybe<Scalars['String']['output']>
  group_id?: Maybe<Scalars['Int']['output']>
  id?: Maybe<Scalars['Int']['output']>
  is_startpage?: Maybe<Scalars['Boolean']['output']>
  lang?: Maybe<Scalars['String']['output']>
  meta_data?: Maybe<Scalars['JsonScalar']['output']>
  name?: Maybe<Scalars['String']['output']>
  parent_id?: Maybe<Scalars['Int']['output']>
  path?: Maybe<Scalars['String']['output']>
  position?: Maybe<Scalars['Int']['output']>
  published_at?: Maybe<Scalars['String']['output']>
  release_id?: Maybe<Scalars['Int']['output']>
  slug?: Maybe<Scalars['String']['output']>
  sort_by_date?: Maybe<Scalars['String']['output']>
  tag_list?: Maybe<Array<Maybe<Scalars['String']['output']>>>
  translated_slugs?: Maybe<Array<Maybe<TranslatedSlug>>>
  uuid?: Maybe<Scalars['String']['output']>
}

export type ContentItems = {
  __typename?: 'ContentItems'
  items?: Maybe<Array<Maybe<ContentItem>>>
  total?: Maybe<Scalars['Int']['output']>
}

export type Datasource = {
  __typename?: 'Datasource'
  id: Scalars['Int']['output']
  name: Scalars['String']['output']
  slug: Scalars['String']['output']
}

export type DatasourceEntries = {
  __typename?: 'DatasourceEntries'
  items: Array<DatasourceEntry>
  total: Scalars['Int']['output']
}

export type DatasourceEntry = {
  __typename?: 'DatasourceEntry'
  dimensionValue?: Maybe<Scalars['String']['output']>
  id: Scalars['Int']['output']
  name: Scalars['String']['output']
  value: Scalars['String']['output']
}

export type Datasources = {
  __typename?: 'Datasources'
  items: Array<Datasource>
}
// ...
export type Link = {
  __typename?: 'Link'
  cachedUrl: Scalars['String']['output']
  email?: Maybe<Scalars['String']['output']>
  fieldtype: Scalars['String']['output']
  id: Scalars['String']['output']
  linktype: Scalars['String']['output']
  story?: Maybe<Story>
  url: Scalars['String']['output']
}

export type LinkEntries = {
  __typename?: 'LinkEntries'
  items: Array<LinkEntry>
}
// ...
export type ProjectComponent = {
  __typename?: 'ProjectComponent'
  _editable?: Maybe<Scalars['String']['output']>
  _uid?: Maybe<Scalars['String']['output']>
  banner?: Maybe<Asset>
  component?: Maybe<Scalars['String']['output']>
  description?: Maybe<Scalars['String']['output']>
  github_link?: Maybe<Link>
  images?: Maybe<Array<Maybe<Asset>>>
  name?: Maybe<Scalars['String']['output']>
  short_description?: Maybe<Scalars['String']['output']>
  stack?: Maybe<Array<Maybe<Scalars['String']['output']>>>
  webiste_link?: Maybe<Link>
  work_in_progress?: Maybe<Scalars['Boolean']['output']>
  year?: Maybe<Scalars['String']['output']>
}

export type ProjectFilterQuery = {
  name?: InputMaybe<FilterQueryOperations>
  short_description?: InputMaybe<FilterQueryOperations>
  stack?: InputMaybe<FilterQueryOperations>
  work_in_progress?: InputMaybe<FilterQueryOperations>
  year?: InputMaybe<FilterQueryOperations>
}

export type ProjectItem = {
  __typename?: 'ProjectItem'
  alternates?: Maybe<Array<Maybe<Alternate>>>
  content?: Maybe<ProjectComponent>
  created_at?: Maybe<Scalars['String']['output']>
  default_full_slug?: Maybe<Scalars['String']['output']>
  first_published_at?: Maybe<Scalars['String']['output']>
  full_slug?: Maybe<Scalars['String']['output']>
  group_id?: Maybe<Scalars['Int']['output']>
  id?: Maybe<Scalars['Int']['output']>
  is_startpage?: Maybe<Scalars['Boolean']['output']>
  lang?: Maybe<Scalars['String']['output']>
  meta_data?: Maybe<Scalars['JsonScalar']['output']>
  name?: Maybe<Scalars['String']['output']>
  parent_id?: Maybe<Scalars['Int']['output']>
  path?: Maybe<Scalars['String']['output']>
  position?: Maybe<Scalars['Int']['output']>
  published_at?: Maybe<Scalars['String']['output']>
  release_id?: Maybe<Scalars['Int']['output']>
  slug?: Maybe<Scalars['String']['output']>
  sort_by_date?: Maybe<Scalars['String']['output']>
  tag_list?: Maybe<Array<Maybe<Scalars['String']['output']>>>
  translated_slugs?: Maybe<Array<Maybe<TranslatedSlug>>>
  uuid?: Maybe<Scalars['String']['output']>
}

export type ProjectItems = {
  __typename?: 'ProjectItems'
  items?: Maybe<Array<Maybe<ProjectItem>>>
  total?: Maybe<Scalars['Int']['output']>
}

export type QueryType = {
  __typename?: 'QueryType'
  ContentNode?: Maybe<ContentItem>
  ContentNodes?: Maybe<ContentItems>
  DatasourceEntries?: Maybe<DatasourceEntries>
  Datasources?: Maybe<Datasources>
  Links?: Maybe<LinkEntries>
  ProjectItem?: Maybe<ProjectItem>
  ProjectItems?: Maybe<ProjectItems>
  RateLimit?: Maybe<RateLimit>
  Space?: Maybe<Space>
  Tags?: Maybe<Tags>
}
// ...
export type RateLimit = {
  __typename?: 'RateLimit'
  maxCost: Scalars['Int']['output']
}

export type Space = {
  __typename?: 'Space'
  domain: Scalars['String']['output']
  id: Scalars['Int']['output']
  languageCodes: Array<Maybe<Scalars['String']['output']>>
  name: Scalars['String']['output']
  version: Scalars['Int']['output']
}
// ...
export type Tags = {
  __typename?: 'Tags'
  items: Array<Tag>
}

export type TranslatedSlug = {
  __typename?: 'TranslatedSlug'
  lang: Scalars['String']['output']
  name?: Maybe<Scalars['String']['output']>
  path?: Maybe<Scalars['String']['output']>
}
// ...
export type GetProjectItemsOverviewQuery = {
  __typename?: 'QueryType'
  ProjectItems?: {
    __typename?: 'ProjectItems'
    items?: Array<{
      __typename?: 'ProjectItem'
      slug?: string | null
      content?: {
        __typename?: 'ProjectComponent'
        name?: string | null
        short_description?: string | null
        banner?: { __typename?: 'Asset'; id?: number | null; filename: string } | null
      } | null
    } | null> | null
  } | null
}